+++ bootstrap.php
@@ -15,8 +15,10 @@ define('LARAVEL_START', microtime(true));
 
 $app = require __DIR__.'/../../../bootstrap/app.php';
 
-$app->make(\Illuminate\Contracts\Console\Kernel::class)
-    ->bootstrap();
+if ($app instanceof \Illuminate\Contracts\Foundation\Application) {
+    $app->make(\Illuminate\Contracts\Console\Kernel::class)
+        ->bootstrap();
+}
 
 $app->make('config')
     ->set('larastan.mixins', require __DIR__.DIRECTORY_SEPARATOR.'config'.DIRECTORY_SEPARATOR.'mixins.php');