+++ src/Console/OptionsResolver.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace NunoMaduro\Larastan\Console;
 
 use PHPStan\Command\AnalyseCommand;
-use Illuminate\Contracts\Foundation\Application;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputDefinition;
 
@@ -34,11 +33,6 @@ final class OptionsResolver
     private const DEFAULT_MEMORY_LIMIT = '2048M';
 
     /**
-     * @var \Illuminate\Contracts\Foundation\Application
-     */
-    private $application;
-
-    /**
      * @var \PHPStan\Command\AnalyseCommand
      */
     private $command;
@@ -51,12 +45,10 @@ final class OptionsResolver
     /**
      * OptionsResolver constructor.
      *
-     * @param \Illuminate\Contracts\Foundation\Application $application
      * @param \PHPStan\Command\AnalyseCommand $command
      */
-    public function __construct(Application $application, AnalyseCommand $command)
+    public function __construct(AnalyseCommand $command)
     {
-        $this->application = $application;
         $this->command = $command;
     }
 
@@ -72,7 +64,7 @@ final class OptionsResolver
             ->setDefault(self::DEFAULT_LEVEL);
 
         $definition->getOption('autoload-file')
-            ->setDefault($this->application->basePath('vendor/autoload.php'));
+            ->setDefault(base_path('vendor/autoload.php'));
 
         $definition->getOption('configuration')
             ->setDefault(__DIR__.'/../../extension.neon');
@@ -82,11 +74,7 @@ final class OptionsResolver
 
         $definition->addOption(
             new InputOption(
-                'paths',
-                'p',
-                InputOption::VALUE_REQUIRED,
-                'Paths with source code to run analysis on',
-                $this->application->make('path')
+                'paths', 'p', InputOption::VALUE_REQUIRED, 'Paths with source code to run analysis on', base_path('app')
             )
         );