+++ src/Methods/Pipes/RedirectResponseWiths.php
@@ -16,20 +16,18 @@ namespace NunoMaduro\Larastan\Methods\Pipes;
 use Closure;
 use function substr;
 use Illuminate\Http\RedirectResponse;
-use NunoMaduro\Larastan\Methods\Passable;
+use NunoMaduro\Larastan\Contracts\Methods\PassableContract;
+use NunoMaduro\Larastan\Contracts\Methods\Pipes\PipeContract;
 
 /**
  * @internal
  */
-final class RedirectResponseWiths
+final class RedirectResponseWiths implements PipeContract
 {
     /**
-     * @param \NunoMaduro\Larastan\Methods\Passable $passable
-     * @param \Closure $next
-     *
-     * @return void
+     * {@inheritdoc}
      */
-    public function handle(Passable $passable, Closure $next): void
+    public function handle(PassableContract $passable, Closure $next): void
     {
         $classReflection = $passable->getClassReflection();
         $methodName = $passable->getMethodName();