+++ src/ReturnTypes/ModelExtension.php
@@ -89,7 +89,7 @@ final class ModelExtension implements DynamicStaticMethodReturnTypeExtension, Br
          * If the method returns a static type, we instruct phpstan that
          * "static" points to the concrete class model.
          */
-        if ($variants[0] instanceof FunctionVariantWithPhpDocs) {
+        if ($methodCall->class instanceof \PhpParser\Node\Name && $variants[0] instanceof FunctionVariantWithPhpDocs) {
             $className = $methodCall->class->toString();
 
             if (class_exists($className)) {
@@ -126,7 +126,7 @@ final class ModelExtension implements DynamicStaticMethodReturnTypeExtension, Br
     {
         foreach ($types as $key => $type) {
             if ($type instanceof ObjectType && $type->getClassName() === Model::class) {
-                unset($types[$key]);
+                $types[$key] = new StaticType($staticType);
             }
 
             if ($type instanceof StaticType) {