+++ CHANGELOG.md
@@ -4,6 +4,57 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.4] - 2018-07-24
+### Fixes
+- Issue while resolving implementations that don't exist on the container
+
+## [0.2.3] - 2018-07-23
+### Fixes
+- Common laravel false positives on trusted proxies and app exception handler
+
+## [0.2.2] - 2018-07-23
+### Fixes
+- Resolved null type from container
+
+## [0.2.1] - 2018-07-23
+### Fixes
+- Error - "internal error: * product does not exist"
+
+## [0.2.0] - 2018-07-22
+### Fixes
+- Issues while using Lumen
+
+## [0.1.9] - 2018-07-22
+### Adds
+- Support to Lumen and Laravel Zero
+
+## [0.1.8] - 2018-07-22
+### Adds
+- Support to builder dynamic wheres
+
+## [0.1.7] - 2018-07-22
+### Adds
+- Support to "object" return type
+
+### Fixes
+- Bug on macro extension
+
+## [0.1.6] - 2018-07-22
+### Adds
+- Allows array access on objects that respects container's contract
+
+## [0.1.5] - 2018-07-22
+### Fixes
+- Removes unused `dd`
+
+## [0.1.4] - 2018-07-22
+### Fixes
+- Issue when `static` is missing was return type hint on Illuminate Model mixins
+
+## [0.1.3] - 2018-07-20
+### Fixes
+- Usage on Windows + Laravel Homestead ([#55](https://github.com/nunomaduro/larastan/pull/55))
+
 ## [0.1.2] - 2018-07-20
 ### Adds
 - `Illuminate\Contracts` property extension